Refactored symbolicator runner
