+++ external-crates/move/crates/move-analyzer/src/completions/mod.rs
@@ -11,7 +11,8 @@ use crate::{
     },
     context::Context,
     symbols::{
-        self, SymbolicatorRunner,
+        self,
+        runner::SymbolicatorRunner,
         types::{CursorContext, PrecomputedPkgInfo, Symbols},
     },
 };
