+++ external-crates/move/crates/move-analyzer/src/inlay_hints.rs
@@ -4,9 +4,9 @@
 use crate::{
     context::Context,
     symbols::{
-        SymbolicatorRunner,
         ide_strings::type_to_ide_string,
         on_hover_markup,
+        runner::SymbolicatorRunner,
         types::{DefInfo, ModuleDefs, Symbols},
     },
 };
